using Switch to set return default page to display error content from other component
