+++ src/components/MenuList/index.js
@@ -1,0 +1,1 @@
+export { default } from './MenuList.js';
