+++ src/components/NotFound/NotFound.css
@@ -1,0 +1,7 @@
+/*
+ * NotFound css File
+*/
+
+.NotFound {
+    /* propertys here */
+}
